fix: refactorizar el codigo para que sea legible
