+++ Saca la Bici/Framework/Views/Actividades/activityCardView.swift
@@ -9,24 +9,27 @@ import SwiftUI
 
 struct ActivityCardView: View {
     var activityTitle: String
+    var activityType: String
     var level: String?
     var date: String?
     var time: String?
     var duration: String?
     var location: String?
     var attendees: Int?
-
-    @State private var isJoined: Bool = false  // Estado para cambiar el botón
+    @ObservedObject private var userSessionManager = UserSessionManager.shared
     
+    @State private var isJoined: Bool = false
+    @State private var isStarted: Bool = false
+
     var body: some View {
-        VStack(alignment: .leading, spacing: 8) { // Añadir espaciado
+        VStack(alignment: .leading, spacing: 8) {
             HStack {
                 Text(activityTitle)
                     .font(.headline)
                     .foregroundColor(.primary)
-                
+
                 Spacer()
-                
+
                 if let level = level {
                     Text(level)
                         .font(.caption)
@@ -34,7 +37,7 @@ struct ActivityCardView: View {
                         .background(level == "Nivel 1" ? Color.green : (level == "Nivel 2" ? Color.orange : Color.gray))
                         .cornerRadius(8)
                 }
-                
+
                 if let attendees = attendees {
                     HStack(spacing: 4) {
                         Image(systemName: "person.2")
@@ -43,7 +46,7 @@ struct ActivityCardView: View {
                     }
                 }
             }
-            
+
             if let date = date {
                 HStack {
                     Text("Fecha")
@@ -54,7 +57,7 @@ struct ActivityCardView: View {
                         .foregroundColor(.primary)
                 }
             }
-            
+
             if let time = time {
                 HStack {
                     Text("Hora")
@@ -65,7 +68,7 @@ struct ActivityCardView: View {
                         .foregroundColor(.primary)
                 }
             }
-            
+
             if let duration = duration {
                 HStack {
                     Text("Duración")
@@ -76,13 +79,12 @@ struct ActivityCardView: View {
                         .foregroundColor(.primary)
                 }
             }
-            
-            // Placeholder for image
+
             Rectangle()
-                .fill(Color.gray)
+                .fill(Color.gray.opacity(0.3))
                 .frame(height: 100)
                 .cornerRadius(8)
-            
+
             if let location = location {
                 HStack {
                     Text("Ubicación")
@@ -93,22 +95,66 @@ struct ActivityCardView: View {
                         .foregroundColor(.primary)
                 }
             }
-            
-            Button {
-                isJoined.toggle()
-            } label: {
-                HStack {
-                    Image(systemName: isJoined ? "xmark.circle" : "plus.circle")
-                    Text(isJoined ? "Cancelar asistencia" : "Unirse")
-                        .font(.headline)
+
+            // Botones según permisos
+            if userSessionManager.puedeIniciarRodada() {
+                VStack(spacing: 8) {
+                    // Botón de iniciar actividad (color verde #88B598)
+                    Button(action: {
+                        isStarted.toggle()
+                    }, label: {
+                        HStack {
+                            Image(systemName: isStarted ? "pause.circle" : "play.circle")
+                            Text(isStarted ? "Parar" : "Iniciar")
+                                .font(.headline)
+                        }
+                        .foregroundColor(.white)
+                        .padding()
+                        .frame(maxWidth: .infinity)
+                        .background(isStarted ? Color.red : ColorManager.shared.colorFromHex("#88B598"))
+                        .cornerRadius(8)
+                    })
+                    
+                    // Botón de asistencia (color amarillo)
+                    Button(action: {
+                        // Lógica de asistencia
+                    }, label: {
+                        HStack {
+                            Image(systemName: "person.circle")
+                                .font(.headline)
+                            Text("Asistencia")
+                        }
+                        .foregroundColor(.white)
+                        .padding()
+                        .frame(maxWidth: .infinity)
+                        .background(Color.yellow)
+                        .cornerRadius(8)
+                    })
                 }
-                .foregroundColor(.white)
-                .padding()
-                .frame(maxWidth: .infinity)
-                .background(isJoined ? Color.red : Color.yellow)
-                .cornerRadius(8)
+            } else {
+                // Botón de unirse
+                Button(action: {
+                    isJoined.toggle()
+                }, label: {
+                    HStack {
+                        Image(systemName: isJoined ? "xmark.circle" : "plus.circle")
+                        Text(isJoined ? "Cancelar asistencia" : "Unirse")
+                            .font(.headline)
+                    }
+                    .foregroundColor(.white)
+                    .padding()
+                    .frame(maxWidth: .infinity)
+                    .background(
+                        isJoined
+                            ? Color.red
+                            : (userSessionManager.puedeIniciarRodada()
+                                ? ColorManager.shared.colorFromHex("#88B598") // Verde si puede iniciar rodada
+                                : Color.yellow) // Amarillo para los demás casos
+                    )
+                    .cornerRadius(8)
+                })
+                .buttonStyle(PlainButtonStyle())
             }
-            .buttonStyle(PlainButtonStyle())
         }
         .padding()
         .background(Color(UIColor.systemBackground))
