+++ Saca la Bici/Framework/Views/Actividades/rodadasView.swift
@@ -9,6 +9,7 @@ import SwiftUI
 
 struct RodadasView: View {
     @StateObject private var viewModel = RodadasViewModel()
+    @ObservedObject private var userSessionManager = UserSessionManager.shared
     
     var body: some View {
         ScrollView {
@@ -23,8 +24,9 @@ struct RodadasView: View {
                     ForEach(viewModel.rodadas) { rodada in
                         ActivityCardView(
                             activityTitle: rodada.actividad.titulo,
+                            activityType: "Rodada", // Tipo de actividad
                             level: rodada.ruta.nivel,
-                            date: formatDate(rodada.actividad.fecha),
+                            date: FechaManager.shared.formatDate(rodada.actividad.fecha),
                             time: rodada.actividad.hora,
                             duration: rodada.actividad.duracion,
                             location: rodada.actividad.ubicacion,
@@ -38,35 +40,5 @@ struct RodadasView: View {
         .onAppear {
             viewModel.fetchRodadas()
         }
-    }
-    
-    func formatDate(_ dateString: String) -> String {
-        let formatter = ISO8601DateFormatter()
-        formatter.formatOptions = [.withInternetDateTime, .withFractionalSeconds]
-        
-        if let date = formatter.date(from: dateString) {
-            return formatDateCustom(date)
-        }
-        
-        formatter.formatOptions = [.withInternetDateTime]
-        if let date = formatter.date(from: dateString) {
-            return formatDateCustom(date)
-        }
-        
-        return dateString
-    }
-    
-    func formatDateCustom(_ date: Date) -> String {
-        let dateFormatter = DateFormatter()
-        dateFormatter.dateStyle = .full
-        dateFormatter.timeStyle = .none
-        dateFormatter.locale = Locale(identifier: "es_MX")
-        return dateFormatter.string(from: date)
-    }
-}
-
-struct RodadasView_Previews: PreviewProvider {
-    static var previews: some View {
-        RodadasView()
     }
 }
